ws server: add http logger and ws logger
