+++ http-server/src/lib.rs
@@ -32,9 +32,6 @@
 
 mod server;
 
-/// Common builders for RPC responses.
-pub mod response;
-
 pub use jsonrpsee_core::server::access_control::{AccessControl, AccessControlBuilder};
 pub use jsonrpsee_core::server::rpc_module::RpcModule;
 pub use jsonrpsee_types as types;
