+++ tests/tests/metrics.rs
@@ -145,7 +145,7 @@ async fn websocket_server(module: RpcModule<()>, counter: Counter) -> Result<(So
 	let server = WsServerBuilder::default()
 		.register_resource("CPU", 6, 2)?
 		.register_resource("MEM", 10, 1)?
-		.set_logger(counter)
+		.set_ws_logger(counter)
 		.build("127.0.0.1:0")
 		.await?;
